frontend/requirements.txt
# frontend/requirements.txt
streamlit==1.32.0
streamlit-lottie==0.0.5
pandas==2.2.0
requests==2.31.0
Pillow==10.2.0
python-dotenv==1.0.0
pdfkit==1.0.0